chore: profile 패키지 분리
