+++ src/main/java/com/debateseason_backend_v1/domain/profile/controller/docs/ProfileControllerV1Docs.java
@@ -1,4 +1,4 @@
-package com.debateseason_backend_v1.domain.user.controller.docs;
+package com.debateseason_backend_v1.domain.profile.controller.docs;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -7,9 +7,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.debateseason_backend_v1.common.response.ApiResult;
 import com.debateseason_backend_v1.common.response.ErrorResponse;
 import com.debateseason_backend_v1.common.response.VoidApiResult;
-import com.debateseason_backend_v1.domain.user.controller.request.ProfileRegisterRequest;
-import com.debateseason_backend_v1.domain.user.controller.request.ProfileUpdateRequest;
-import com.debateseason_backend_v1.domain.user.service.response.ProfileResponse;
+import com.debateseason_backend_v1.domain.profile.controller.request.ProfileRegisterRequest;
+import com.debateseason_backend_v1.domain.profile.controller.request.ProfileUpdateRequest;
+import com.debateseason_backend_v1.domain.profile.controller.request.request.ProfileRegisterRequest;
+import com.debateseason_backend_v1.domain.profile.controller.request.request.ProfileUpdateRequest;
+import com.debateseason_backend_v1.domain.profile.service.response.ProfileResponse;
 import com.debateseason_backend_v1.security.CustomUserDetails;
 
 import io.swagger.v3.oas.annotations.Operation;
