+++ src/main/java/com/debateseason_backend_v1/domain/profile/controller/request/ProfileRegisterRequest.java
@@ -1,8 +1,8 @@
-package com.debateseason_backend_v1.domain.user.controller.request;
+package com.debateseason_backend_v1.domain.profile.controller.request.request;
 
-import com.debateseason_backend_v1.domain.user.enums.AgeRangeType;
-import com.debateseason_backend_v1.domain.user.enums.GenderType;
-import com.debateseason_backend_v1.domain.user.service.request.ProfileRegisterServiceRequest;
+import com.debateseason_backend_v1.domain.profile.enums.AgeRangeType;
+import com.debateseason_backend_v1.domain.profile.enums.GenderType;
+import com.debateseason_backend_v1.domain.profile.service.request.ProfileRegisterServiceRequest;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
