+++ src/main/java/com/debateseason_backend_v1/domain/profile/service/ProfileServiceV1.java
@@ -1,4 +1,4 @@
-package com.debateseason_backend_v1.domain.user.service;
+package com.debateseason_backend_v1.domain.profile.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,11 +11,11 @@ import com.debateseason_backend_v1.domain.repository.ProfileRepository;
 import com.debateseason_backend_v1.domain.repository.entity.Community;
 import com.debateseason_backend_v1.domain.repository.entity.Profile;
 import com.debateseason_backend_v1.domain.repository.entity.ProfileCommunity;
-import com.debateseason_backend_v1.domain.user.service.request.ProfileRegisterServiceRequest;
-import com.debateseason_backend_v1.domain.user.service.request.ProfileUpdateServiceRequest;
-import com.debateseason_backend_v1.domain.user.service.response.ProfileResponse;
-import com.debateseason_backend_v1.domain.user.validator.CommunityValidator;
-import com.debateseason_backend_v1.domain.user.validator.ProfileValidator;
+import com.debateseason_backend_v1.domain.profile.service.request.ProfileRegisterServiceRequest;
+import com.debateseason_backend_v1.domain.profile.service.request.ProfileUpdateServiceRequest;
+import com.debateseason_backend_v1.domain.profile.service.response.ProfileResponse;
+import com.debateseason_backend_v1.domain.profile.validator.CommunityValidator;
+import com.debateseason_backend_v1.domain.profile.validator.ProfileValidator;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
